requirements.txt
yfinance
pandas
numpy
scikit-learn
joblib
streamlit
matplotlib
ta
python-dotenv
bcrypt
twilio
requests
